fitpack interface is now linked and usable as module in elmer fortan solver/function. Butchered the Makefile a bit to get this done, but things seem to be stable enough for now

+++ include/fitpack/fitpack_interface.f90
@@ -36,13 +36,13 @@ module fitpack_interface
       !   ier  : error flag
       !     ier = 0 : normal return
       !     ier =10 : invalid input data (see restrictions)
-      real, dimension(n), intent(in)  :: t
+      real(8), dimension(n), intent(in)  :: t
       integer, intent(in)  :: n
-      real, dimension(n), intent(in)  :: c
+      real(8), dimension(n), intent(in)  :: c
       integer, intent(in)  :: k
-      real, dimension(n), intent(in)  :: x
+      real(8), dimension(n), intent(in)  :: x
       integer, intent(in)  :: m
-      real, dimension(m), intent(out) :: y
+      real(8), dimension(m), intent(out) :: y
       integer, intent(out) :: ier
     end subroutine splev
